Add long type

+++ src/Main.java
@@ -2,13 +2,14 @@ package src;
 
 public class Main {
     public static void main(String[] args) {
-        int[] salesData = {5, 12, 20, 8, 15, 30, 12, 18};
+        long[] salesData = {5, 12, 20, 8, 15, 30, 12, 18};
         SalesManager salesManager = new SalesManager(salesData);
 
-        int maxSale = salesManager.max();
+        long maxSale = salesManager.max();
         System.out.println("The maximum sale is: " + maxSale);
 
         double trimmedMean = salesManager.calculateTrimmedMean();
         System.out.println("Trimmed mean: " + trimmedMean);
+
     }
 }
